% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_plots.R
\name{combined_plots}
\alias{combined_plots}
\title{Graphically analyse through multiple plots a call to \code{drawsneeded()}}
\usage{
combined_plots(
  expected_error_rate,
  allowed_error_rate,
  cert = 0.95,
  max_n = 1000,
  S = 10000,
  min_prob = 1.5
)
}
\arguments{
\item{expected_error_rate}{The estimated error rate from earlier knowledge.}

\item{allowed_error_rate}{The highest error rate that is still acceptable.
Should be higher than expected_error_rate.}

\item{cert}{The certainty level you want, e.g. \code{0.95}.}

\item{max_n}{The maximum number of samples you are willing to use.}

\item{S}{The number of points on the X-axis of the
plot.}

\item{min_prob}{Only points with at least that probability are
shown in the first plot.}
}
\value{
A ggplot.

A ggplot.
}
\description{
Show in one figure the result of a call to \code{drawsneeded()} by means of four plots.
The first plot shows what chance graph will result if the foreseen number of
errors do indeed emerge by drawing and checking the planned number of
monetary units.
}
\details{
The other three plots analyse the influence of respectively the parameters
\code{expected_error_rate}, \code{allowed_error_rate} and \code{cert} on the number of
draws needed. "Draws needed": meaning here the number of draws needed to have
enough random monetary units seen, to conclude with certainty cert
that the error rate in the monetary unit lies below the allowed_error_rate.
}
\examples{
  combined_plots(0.001, 0.02, cert = 0.95, max_n = 500)
}
