% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_trailing_zeros.R
\name{formatf_without_trailing_zeros}
\alias{formatf_without_trailing_zeros}
\title{print number in fixed point decimal notation, but no trailing zeros}
\usage{
formatf_without_trailing_zeros(nr)
}
\arguments{
\item{nr}{The input number.}
}
\value{
The input number, as a string in fixed point decimal notation,
  with trailing zeros removed.

examples
  formatf_without_trailing_zeros(37.5600)
}
\description{
print number in fixed point decimal notation, but no trailing zeros
}
