% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_plot_varying_allowed_defect_rate.R
\name{margin_plot_varying_allowed_defect_rate}
\alias{margin_plot_varying_allowed_defect_rate}
\title{Show in plot how allowed defect rate influences draws needed}
\usage{
margin_plot_varying_allowed_defect_rate(
  posited_defect_rate = 0.01,
  cert = 0.95,
  max_n = 1000,
  S = 10000
)
}
\arguments{
\item{posited_defect_rate}{The estimated defect rate from earlier knowledge.}

\item{cert}{The certainty level you want, e.g. \code{0.95}.}

\item{max_n}{The number of to be drawn items maximally shown in the plot.}

\item{S}{The number of points on the X-axis of the
plot.}
}
\value{
A ggplot.
}
\description{
The plot is based on varying the variable
allowed_defect_rate in a call to drawsneeded(posited_defect_rate, allowed_defect_rate, cert)
}
\details{
To keep the plot interesting, only the points representing up to
max_n draws are shown.
}
\examples{
  margin_plot_varying_allowed_defect_rate(posited_defect_rate = 0.01, cert = 0.95)
}
