% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_varying_cert.R
\name{plot_varying_cert}
\alias{plot_varying_cert}
\title{Show in plot how varying allowed defect rate influences draws needed}
\usage{
plot_varying_cert(
  posited_defect_rate = 0.01,
  allowed_defect_rate = 0.02,
  max_n = 1000,
  S = 10000
)
}
\arguments{
\item{posited_defect_rate}{The proposed defect rate.
Should be lower than allowed_defect_rate.}

\item{allowed_defect_rate}{The highest defect rate that is still acceptable.
Should be higher than posited_defect_rate.}

\item{max_n}{The number of to be drawn items maximally shown in the plot.#'}

\item{S}{The number of points on the X-axis of the
plot.}
}
\value{
A ggplot.
}
\description{
The plot is based on varying the variable
allowed_defect_rate in a call to drawsneeded(posited_defect_rate, allowed_defect_rate, cert).
}
\details{
To keep the plot interesting, only the points representing up to
max_n draws are shown.
}
\examples{
  plot_varying_cert(posited_defect_rate = 0.01, allowed_defect_rate = 0.02)
}
