% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_trailing_zeros.R
\name{remove_trailing_zeros}
\alias{remove_trailing_zeros}
\title{remove trailing zeros from string containing at least one "."}
\usage{
remove_trailing_zeros(s)
}
\arguments{
\item{s}{The input string.}
}
\value{
The input string with trailing zeros, after ".", if any, removed.

examples
  remove_trailing_zeros("37.5600")
}
\description{
Also removes "." if that is the last character.
The idea is that we trim zeros from a number containing decimal places.
}
