% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_plots.R
\name{combined_plots}
\alias{combined_plots}
\title{Graphically analyse through multiple plots a call to \code{drawsneeded()}}
\usage{
combined_plots(
  posited_defect_rate,
  allowed_defect_rate,
  cert = 0.95,
  max_n = 1000,
  S = 10000,
  high_density_area = 0.999
)
}
\arguments{
\item{posited_defect_rate}{The estimated defect rate from earlier knowledge.}

\item{allowed_defect_rate}{The highest defect rate that is still acceptable.
Should be higher than posited_defect_rate.}

\item{cert}{The certainty level you want, e.g. \code{0.95}.}

\item{max_n}{The number of to be drawn items maximally shown in the plot.}

\item{S}{The number of points on the X-axis of the
plot.}

\item{high_density_area}{Only points with a probability high enough to
be part of the high density interval, which has size high_density_area,
are shown in the plot.}
}
\value{
A ggplot.

A ggplot.
}
\description{
Show in one figure the result of a call to \code{drawsneeded()} by means of four plots.
The first plot shows what chance graph will result if the foreseen number of
defects do indeed emerge by drawing and checking the planned number of
monetary units.
}
\details{
The other three plots analyse the influence of respectively the parameters
\code{posited_defect_rate}, \code{allowed_defect_rate} and \code{cert} on the number of
draws needed. "Draws needed": meaning here the number of draws needed to have
enough random monetary units seen, to conclude with certainty cert
that the defect rate in the monetary unit lies below the allowed_defect_rate.
}
\examples{
  combined_plots(0.001, 0.02, cert = 0.95)
}
