% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsneeded.R
\name{drawsneeded}
\alias{drawsneeded}
\title{compute number of samples needed to establish maximum defect rate with enough certainty}
\usage{
drawsneeded(posited_defect_rate, allowed_defect_rate, cert = 0.95)
}
\arguments{
\item{posited_defect_rate}{The defect rate of which we want to see how many
samples it would take to establish that it is below the allowed
defect rate.
Could for example be our guess of the defect rate.}

\item{allowed_defect_rate}{The highest defect rate that is still acceptable.
So the threshold.
Should be higher than posited_defect_rate.}

\item{cert}{The certainty level you want, e.g. \code{0.95}.}
}
\value{
An estimate of the needed number of samples. If the number of
    needed samples comes to close to the maximum R integer value, return -1.
}
\description{
In sampling we have a file of like items.
We want to estimate the overall defect rate of  the items.
This function guesses the number of draws needed
to establish with some certainty that the defect rate is below a certain
threshold.
}
\details{
Each of the three arguments can have length > 1, but only one of these
three arguments.

Note that the prior used by drawsneeded() is flat: each possible defect rate
is beforehand given an equal probability.
}
\examples{
  x <- drawsneeded(0.001, 0.02, cert = 0.95)
}
