% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsneeded_plot.R
\name{drawsneeded_plot}
\alias{drawsneeded_plot}
\title{Graphically show the results of a call to \code{drawsneeded()}}
\usage{
drawsneeded_plot(
  posited_defect_rate,
  allowed_defect_rate,
  cert = 0.95,
  S = 1e+05,
  min_prob = 1.5
)
}
\arguments{
\item{posited_defect_rate}{The estimated defect rate from earlier knowledge.}

\item{allowed_defect_rate}{The highest defect rate that is still acceptable.
Should be higher than posited_defect_rate.}

\item{cert}{The certainty level you want, e.g. \code{0.95}.#'}

\item{S}{The number of points on the X-axis of the
plot. but less points might be shown, as only points
with a probability density >= min_prob are shown.}

\item{min_prob}{Only points with at least that probability are
shown in the plot.}
}
\value{
A ggplot.

A ggplot.
}
\description{
Let n be the result of the call
  n <- drawsneeded(posited_defect_rate = eer,
                   allowed_defect_rate = aer,
                   cert = c).
chance density graph for k = n*eer, n = n.
The vertical lines to denote posited_defect_rate and allowed_defect_rate are
also on the graph.
}
\details{
If n < 1, show a diagnostic message.

There is no support for vector args with length > 1.
}
\examples{
  drawsneeded_plot(0.001, 0.02, cert = 0.95)
}
